requirements.txt
matrix-client >= 0.3.2
pyTeamUp >= 0.1.1a0
pandas >= 0.24.2
google-api-python-client >= 1.7.11
google-auth-httplib2 >= 0.0.3
google-auth-oauthlib >= 0.4.0

